Serializing grids
