+++ include/jet/scalar_grid3.h
@@ -132,12 +132,6 @@ class ScalarGrid3 : public ScalarField3, public Grid3 {
     void parallelForEachDataPointIndex(
         const std::function<void(size_t, size_t, size_t)>& func) const;
 
-    //! Serializes the grid instance to the output stream \p strm.
-    void serialize(std::ostream* strm) const override;
-
-    //! Deserializes the input stream \p strm to the grid instance.
-    void deserialize(std::istream* strm) override;
-
     // ScalarField3 implementations
 
     //!
@@ -161,6 +155,12 @@ class ScalarGrid3 : public ScalarField3, public Grid3 {
 
     //! Returns the Laplacian at given position \p x.
     double laplacian(const Vector3D& x) const override;
+
+    //! Serializes the grid instance to the output buffer.
+    void serialize(std::vector<uint8_t>* buffer) const override;
+
+    //! Deserializes the input buffer to the grid instance.
+    void deserialize(const std::vector<uint8_t>& buffer) override;
 
     //! Fetches the data into a continuous linear array.
     void getData(std::vector<double>* data) const override;
