+++ src/jet/CPPLINT.cfg
@@ -3,5 +3,9 @@ filter=-readability/fn_size
 exclude_files=marching_cubes_table\.h
 exclude_files=pch\.h
 exclude_files=basic_types_generated.h
+exclude_files=scalar_grid2_generated.h
+exclude_files=scalar_grid3_generated.h
+exclude_files=vector_grid2_generated.h
+exclude_files=vector_grid3_generated.h
 exclude_files=grid_system_data2_generated.h
 exclude_files=grid_system_data3_generated.h
