+++ src/jet/grid3.cpp
@@ -4,7 +4,6 @@
 #include <jet/grid3.h>
 #include <jet/parallel.h>
 #include <jet/serial.h>
-
 #include <algorithm>
 #include <fstream>
 #include <iostream>
@@ -106,36 +105,4 @@ void Grid3::setGrid(const Grid3& other) {
     _gridSpacing = other._gridSpacing;
     _origin = other._origin;
     _boundingBox = other._boundingBox;
-}
-
-void Grid3::serializeGrid(std::ostream* strm) const {
-    uint64_t res64[3] = { _resolution.x, _resolution.y, _resolution.z };
-
-    const char* resAsBytes = reinterpret_cast<const char*>(&res64);
-    const char* gsAsBytes = reinterpret_cast<const char*>(&_gridSpacing);
-    const char* orgAsBytes = reinterpret_cast<const char*>(&_origin);
-    const char* boxAsBytes = reinterpret_cast<const char*>(&_boundingBox);
-
-    strm->write(resAsBytes, 3 * sizeof(uint64_t));
-    strm->write(gsAsBytes, 3 * sizeof(double));
-    strm->write(orgAsBytes, 3 * sizeof(double));
-    strm->write(boxAsBytes, 6 * sizeof(double));
-}
-
-void Grid3::deserializeGrid(std::istream* strm) {
-    uint64_t res64[3];
-    char* resAsBytes = reinterpret_cast<char*>(res64);
-    char* gsAsBytes = reinterpret_cast<char*>(&_gridSpacing);
-    char* orgAsBytes = reinterpret_cast<char*>(&_origin);
-    char* boxAsBytes = reinterpret_cast<char*>(&_boundingBox);
-
-    strm->read(resAsBytes, 3 * sizeof(uint64_t));
-    strm->read(gsAsBytes, 3 * sizeof(double));
-    strm->read(orgAsBytes, 3 * sizeof(double));
-    strm->read(boxAsBytes, 6 * sizeof(double));
-
-    _resolution = Size3(
-        static_cast<size_t>(res64[0]),
-        static_cast<size_t>(res64[1]),
-        static_cast<size_t>(res64[2]));
 }
