+++ src/jet/grid_system_data3.fbs
@@ -1,22 +1,8 @@
 include "basic_types.fbs";
+include "scalar_grid3.fbs";
+include "vector_grid3.fbs";
 
 namespace jet.fbs;
-
-table ScalarGrid3 {
-    type:string;
-    resolution:Size3;
-    gridSpacing:Vector3D;
-    origin:Vector3D;
-    data:[double];
-}
-
-table VectorGrid3 {
-    type:string;
-    resolution:Size3;
-    gridSpacing:Vector3D;
-    origin:Vector3D;
-    data:[double];
-}
 
 table GridSystemData3 {
     resolution:Size3;
