+++ src/jet/scalar_grid2.fbs
@@ -1,0 +1,13 @@
+include "basic_types.fbs";
+
+namespace jet.fbs;
+
+table ScalarGrid2 {
+    type:string;
+    resolution:Size2;
+    gridSpacing:Vector2D;
+    origin:Vector2D;
+    data:[double];
+}
+
+root_type ScalarGrid2;
