+++ src/jet/scalar_grid3.fbs
@@ -1,0 +1,13 @@
+include "basic_types.fbs";
+
+namespace jet.fbs;
+
+table ScalarGrid3 {
+    type:string;
+    resolution:Size3;
+    gridSpacing:Vector3D;
+    origin:Vector3D;
+    data:[double];
+}
+
+root_type ScalarGrid3;
