+++ src/jet/vector_grid3_generated.h
@@ -1,0 +1,90 @@
+// automatically generated by the FlatBuffers compiler, do not modify
+
+#ifndef FLATBUFFERS_GENERATED_VECTORGRID3_JET_FBS_H_
+#define FLATBUFFERS_GENERATED_VECTORGRID3_JET_FBS_H_
+
+#include "flatbuffers/flatbuffers.h"
+
+#include "basic_types_generated.h"
+
+namespace jet {
+namespace fbs {
+
+struct VectorGrid3;
+
+struct VectorGrid3 FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  enum {
+    VT_TYPE = 4,
+    VT_RESOLUTION = 6,
+    VT_GRIDSPACING = 8,
+    VT_ORIGIN = 10,
+    VT_DATA = 12
+  };
+  const flatbuffers::String *type() const { return GetPointer<const flatbuffers::String *>(VT_TYPE); }
+  const jet::fbs::Size3 *resolution() const { return GetStruct<const jet::fbs::Size3 *>(VT_RESOLUTION); }
+  const jet::fbs::Vector3D *gridSpacing() const { return GetStruct<const jet::fbs::Vector3D *>(VT_GRIDSPACING); }
+  const jet::fbs::Vector3D *origin() const { return GetStruct<const jet::fbs::Vector3D *>(VT_ORIGIN); }
+  const flatbuffers::Vector<double> *data() const { return GetPointer<const flatbuffers::Vector<double> *>(VT_DATA); }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyField<flatbuffers::uoffset_t>(verifier, VT_TYPE) &&
+           verifier.Verify(type()) &&
+           VerifyField<jet::fbs::Size3>(verifier, VT_RESOLUTION) &&
+           VerifyField<jet::fbs::Vector3D>(verifier, VT_GRIDSPACING) &&
+           VerifyField<jet::fbs::Vector3D>(verifier, VT_ORIGIN) &&
+           VerifyField<flatbuffers::uoffset_t>(verifier, VT_DATA) &&
+           verifier.Verify(data()) &&
+           verifier.EndTable();
+  }
+};
+
+struct VectorGrid3Builder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_type(flatbuffers::Offset<flatbuffers::String> type) { fbb_.AddOffset(VectorGrid3::VT_TYPE, type); }
+  void add_resolution(const jet::fbs::Size3 *resolution) { fbb_.AddStruct(VectorGrid3::VT_RESOLUTION, resolution); }
+  void add_gridSpacing(const jet::fbs::Vector3D *gridSpacing) { fbb_.AddStruct(VectorGrid3::VT_GRIDSPACING, gridSpacing); }
+  void add_origin(const jet::fbs::Vector3D *origin) { fbb_.AddStruct(VectorGrid3::VT_ORIGIN, origin); }
+  void add_data(flatbuffers::Offset<flatbuffers::Vector<double>> data) { fbb_.AddOffset(VectorGrid3::VT_DATA, data); }
+  VectorGrid3Builder(flatbuffers::FlatBufferBuilder &_fbb) : fbb_(_fbb) { start_ = fbb_.StartTable(); }
+  VectorGrid3Builder &operator=(const VectorGrid3Builder &);
+  flatbuffers::Offset<VectorGrid3> Finish() {
+    auto o = flatbuffers::Offset<VectorGrid3>(fbb_.EndTable(start_, 5));
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<VectorGrid3> CreateVectorGrid3(flatbuffers::FlatBufferBuilder &_fbb,
+    flatbuffers::Offset<flatbuffers::String> type = 0,
+    const jet::fbs::Size3 *resolution = 0,
+    const jet::fbs::Vector3D *gridSpacing = 0,
+    const jet::fbs::Vector3D *origin = 0,
+    flatbuffers::Offset<flatbuffers::Vector<double>> data = 0) {
+  VectorGrid3Builder builder_(_fbb);
+  builder_.add_data(data);
+  builder_.add_origin(origin);
+  builder_.add_gridSpacing(gridSpacing);
+  builder_.add_resolution(resolution);
+  builder_.add_type(type);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<VectorGrid3> CreateVectorGrid3Direct(flatbuffers::FlatBufferBuilder &_fbb,
+    const char *type = nullptr,
+    const jet::fbs::Size3 *resolution = 0,
+    const jet::fbs::Vector3D *gridSpacing = 0,
+    const jet::fbs::Vector3D *origin = 0,
+    const std::vector<double> *data = nullptr) {
+  return CreateVectorGrid3(_fbb, type ? _fbb.CreateString(type) : 0, resolution, gridSpacing, origin, data ? _fbb.CreateVector<double>(*data) : 0);
+}
+
+inline const jet::fbs::VectorGrid3 *GetVectorGrid3(const void *buf) { return flatbuffers::GetRoot<jet::fbs::VectorGrid3>(buf); }
+
+inline bool VerifyVectorGrid3Buffer(flatbuffers::Verifier &verifier) { return verifier.VerifyBuffer<jet::fbs::VectorGrid3>(nullptr); }
+
+inline void FinishVectorGrid3Buffer(flatbuffers::FlatBufferBuilder &fbb, flatbuffers::Offset<jet::fbs::VectorGrid3> root) { fbb.Finish(root); }
+
+}  // namespace fbs
+}  // namespace jet
+
+#endif  // FLATBUFFERS_GENERATED_VECTORGRID3_JET_FBS_H_
